ISSUE/TEAM/002-Migliorata gestione .env.dist durante setup.sh, rimosso codice superfluo setup.sh, aggiunte variabili al .env.dist

+++ setup.sh
@@ -5,9 +5,6 @@ provisioning_dir="${PWD}"
 source_dir="${provisioning_dir}/Docker/${app_name}/source"
 repo_source="ssh://git-codecommit.eu-west-1.amazonaws.com/v1/repos/tb_taste_purchase"
 
-export DOCKER_UID="$(id -u)"
-export DOCKER_GID="$(id -g)"
-
 (
   set -o errexit
   set -o pipefail
@@ -15,10 +12,18 @@ export DOCKER_GID="$(id -g)"
 
   echo "Installing application..."
 
-  cp -p "${provisioning_dir}/.env.dist" "${provisioning_dir}/.env"
+  if [ ! -f "${provisioning_dir}/.env" ]
+  then
+    read -n 1 -s -r -p "File .env does not exist. Do you want to copy it from .env.dist? (y/n)" response_copy
+    echo ""
+    if [[ "${response_copy}" == "y" || "${response_copy}" == "Y" ]]
+    then
+      cp -p "${provisioning_dir}/.env.dist" "${provisioning_dir}/.env"
+    fi
+  fi
 
-  echo "Edit the .env file with your variables"
-  read -n 1 -s -r -p "Press any key to continue when you are finished"
+  echo "Please, check variables in the .env file."
+  read -n 1 -s -r -p "Press any key to continue when you are sure every variable is correct."
   echo ""
 
   set -o allexport
@@ -42,10 +47,3 @@ export DOCKER_GID="$(id -g)"
 
   xdg-open "http://${HTTP_HOST}:${HTTP_PORT}"
 )
-
-if [ $? -eq 0 ]
-then
-  echo "Install successfull"
-else
-  return $?
-fi
